Move Menu components into single navigation folder
