+++ src/pages/biography.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { graphql } from 'gatsby';
 import Template from '../layouts/Template';
-import Menu from '../components/Menu';
+import Menu from '../components/navigation/Menu';
 import Body from '../components/Body';
 import BiographyItem from '../components/biography/BiographyItem';
